chore(browser): render error pages within webview (#621)
* feat(browser): render error pages within render

* chore(preload): split preload across files

* feat(browser): console.error when file can't be loaded
