+++ __tests__/renderer/browser/index.test.js
@@ -1,13 +1,11 @@
 import React from 'react';
 import { mount } from 'enzyme';
-import { merge } from 'lodash';
 
 import { provideStore, createStore } from 'testHelpers';
 
 import { Browser } from 'browser';
 import { EXTERNAL } from 'browser/values/browserValues';
 import DAppContainer from 'browser/components/DAppContainer';
-import Error from 'browser/components/Error';
 
 const initialState = {
   browser: {
@@ -20,9 +18,7 @@ const initialState = {
         icon: 'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAADUlEQVR42mNk+P+/HgAFhAJ/wlseKgAAAABJRU5ErkJggg==',
         addressBarEntry: true,
         loading: false,
-        requestCount: 1,
-        errorCode: null,
-        errorDescription: null
+        requestCount: 1
       },
       'tab-2': {
         type: EXTERNAL,
@@ -31,9 +27,7 @@ const initialState = {
         icon: null,
         addressBarEntry: true,
         loading: false,
-        requestCount: 1,
-        errorCode: null,
-        errorDescription: null
+        requestCount: 1
       }
     }
   }
@@ -54,29 +48,5 @@ describe('<Browser />', () => {
     const wrapper = mountBrowser();
     expect(wrapper.find(DAppContainer).at(0).hasClass('active')).toBe(true);
     expect(wrapper.find(DAppContainer).at(1).hasClass('active')).toBe(false);
-  });
-
-  describe('when a tab has an error', () => {
-    const state = merge({}, initialState, {
-      browser: {
-        tabs: {
-          'tab-3': {
-            type: EXTERNAL,
-            target: 'protocol://error',
-            title: 'nOS Error',
-            addressBarEntry: true,
-            loading: false,
-            requestCount: 1,
-            errorCode: -105,
-            errorDescription: 'NAME_NOT_RESOLVED'
-          }
-        }
-      }
-    });
-
-    it('hides the webview and shows an error', () => {
-      const wrapper = mountBrowser(state);
-      expect(wrapper.find(Error).length).toBe(1);
-    });
   });
 });
