+++ src/renderer/browser/components/DAppContainer/DAppContainer.scss
@@ -4,9 +4,5 @@
 
   .webview {
     height: 100%;
-
-    &.hidden {
-      display: none;
-    }
   }
 }
