--- src/renderer/browser/components/Error/Error.js
@@ -1,38 +1,0 @@
-import React from 'react';
-import { string, number } from 'prop-types';
-
-import styles from './Error.scss';
-
-export default class Error extends React.PureComponent {
-  static propTypes = {
-    target: string.isRequired,
-    code: number.isRequired,
-    description: string.isRequired
-  }
-
-  render() {
-    return (
-      <div className={styles.error}>
-        {this.renderTitle()}
-        {this.renderDescription()}
-      </div>
-    );
-  }
-
-  renderTitle() {
-    return (
-      <h1>Error {this.props.code}</h1>
-    );
-  }
-
-  renderDescription() {
-    return (
-      <div>
-        <p>
-          <strong>{this.props.target}</strong> could not be found.
-        </p>
-        <p>{this.props.description}</p>
-      </div>
-    );
-  }
-}
