--- src/renderer/browser/components/Error/Error.scss
@@ -1,7 +1,0 @@
-.error {
-  display: flex;
-  height: 100%;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-}
